big api update

+++ API/Controllers/Company/CompanyController.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Web.Http;
 using Domain.Requests;
 using LOGIC.Services.Company;
@@ -8,17 +9,19 @@ namespace API.Controllers.Company
     public class CompanyController : ApiController
     {
         private readonly CompanyService _companyService;
-        
+
         public CompanyController()
         {
             _companyService = new CompanyService();
         }
-        
+
         [HttpGet]
         [Route("companies")]
         public IHttpActionResult GetAll()
         {
-            var results = _companyService.GetAll(1, 1);
+            var results = _companyService.GetAll();
+            if (results.ToArray().Length == 0) return Ok("There are no companies found");
+
             return Ok(results);
         }
 
@@ -33,17 +36,30 @@ namespace API.Controllers.Company
 
         [HttpPost]
         [Route("companies")]
-        public IHttpActionResult Create(CompanyRequest entity)
+        public IHttpActionResult Create([FromBody] CompanyRequest companyToCreate)
         {
-            var result = _companyService.Create(entity);
+            if (companyToCreate == null) return BadRequest("Empty request body!");
+            if (companyToCreate.Name == null)
+                return BadRequest("You cannot create a company without specifying the name");
+            if (companyToCreate.Description == null)
+                return BadRequest("You cannot create a company without specifying the description");
+
+            var temp = _companyService.GetByName(companyToCreate.Name);
+
+            if (temp != null) return BadRequest("A company with that name already exists!");
+
+            var result = _companyService.Create(companyToCreate);
             return Ok(result);
         }
 
         [HttpPut]
         [Route("companies/{id}")]
-        public IHttpActionResult Update(int id, CompanyRequest entity)
+        public IHttpActionResult Update(int id, CompanyRequest companyRequest)
         {
-            var result = _companyService.Update(id, entity);
+            var company = _companyService.GetById(id);
+            if (company == null) return NotFound();
+
+            var result = _companyService.Update(id, companyRequest);
             return Ok(result);
         }
 
@@ -51,6 +67,8 @@ namespace API.Controllers.Company
         [Route("companies/{id}")]
         public IHttpActionResult Delete(int id)
         {
+            var company = _companyService.GetById(id);
+            if (company == null) return NotFound();
             var result = _companyService.Delete(id);
             return Ok(result);
         }
